add more unit-tests

+++ src/main/java/org/python/integration/object/PythonInt.java
@@ -4,6 +4,7 @@ package org.python.integration.object;
 import java.math.BigInteger;
 
 public class PythonInt extends AbstractPythonObject {
+
     private PythonInt(long index, long scopeId) {
         super(index, scopeId);
     }
@@ -14,7 +15,7 @@ public class PythonInt extends AbstractPythonObject {
     }
 
     public int toJavaInt() {
-        return (int) this.toJavaLong();
+        return Math.toIntExact(this.toJavaLong());
     }
 
     public native long toJavaLong();
